// frontend/src/api-client/models/TicketListItemDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TicketTracking.Api | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PriorityDto2 } from './PriorityDto2';
import {
    PriorityDto2FromJSON,
    PriorityDto2FromJSONTyped,
    PriorityDto2ToJSON,
    PriorityDto2ToJSONTyped,
} from './PriorityDto2';
import type { TicketTypeDto2 } from './TicketTypeDto2';
import {
    TicketTypeDto2FromJSON,
    TicketTypeDto2FromJSONTyped,
    TicketTypeDto2ToJSON,
    TicketTypeDto2ToJSONTyped,
} from './TicketTypeDto2';
import type { StatusDto2 } from './StatusDto2';
import {
    StatusDto2FromJSON,
    StatusDto2FromJSONTyped,
    StatusDto2ToJSON,
    StatusDto2ToJSONTyped,
} from './StatusDto2';

/**
 * 
 * @export
 * @interface TicketListItemDto
 */
export interface TicketListItemDto {
    /**
     * 
     * @type {number}
     * @memberof TicketListItemDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof TicketListItemDto
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TicketListItemDto
     */
    applicationName?: string | null;
    /**
     * 
     * @type {PriorityDto2}
     * @memberof TicketListItemDto
     */
    priority?: PriorityDto2 | null;
    /**
     * 
     * @type {TicketTypeDto2}
     * @memberof TicketListItemDto
     */
    ticketType?: TicketTypeDto2 | null;
    /**
     * 
     * @type {StatusDto2}
     * @memberof TicketListItemDto
     */
    status?: StatusDto2 | null;
}

/**
 * Check if a given object implements the TicketListItemDto interface.
 */
export function instanceOfTicketListItemDto(value: object): value is TicketListItemDto {
    return true;
}

export function TicketListItemDtoFromJSON(json: any): TicketListItemDto {
    return TicketListItemDtoFromJSONTyped(json, false);
}

export function TicketListItemDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TicketListItemDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
        'applicationName': json['applicationName'] == null ? undefined : json['applicationName'],
        'priority': json['priority'] == null ? undefined : PriorityDto2FromJSON(json['priority']),
        'ticketType': json['ticketType'] == null ? undefined : TicketTypeDto2FromJSON(json['ticketType']),
        'status': json['status'] == null ? undefined : StatusDto2FromJSON(json['status']),
    };
}

export function TicketListItemDtoToJSON(json: any): TicketListItemDto {
    return TicketListItemDtoToJSONTyped(json, false);
}

export function TicketListItemDtoToJSONTyped(value?: TicketListItemDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'applicationName': value['applicationName'],
        'priority': PriorityDto2ToJSON(value['priority']),
        'ticketType': TicketTypeDto2ToJSON(value['ticketType']),
        'status': StatusDto2ToJSON(value['status']),
    };
}